only fail on error nodes when using a "--strict" flag
Tree-sitter is error-tolerant, so it's probably a mistake to abort every
time an error is encountered.

Trying to define which error-states can be handled as was previously
done for resource-like class definitions isn't really viable, and I
haven't seen any error nodes that lead to breaking the code so far.