switch grammar to https://github.com/smoeding/tree-sitter-puppet
There are almost never any error nodes after this, so when they now
occur no attempt is made to format them, instead the remaining lines are
passed unchanged.

The grammar has support for heredocs, so all heredoc-related workarounds
have been removed.

Move to BSD 2-Clause license to match the one of the grammar.